Add files via upload
I am a junior front end developer, still learning the ropes after 2 years without practice due to 2020 events such as the pandemic, lack  of motivation and other mental issues that needed to be dealt with. I finished my studies in Dec 2021 and had no luck finding internships in 2022 so i spent the time trying to practice and this was amongst the first projects I made in my relearning experience still want to learn more and to anyone that sees this help and constructive feedback is appreciated.